refactor(benches): consolidate benchmark suite from 12 to 9 files
## Summary
Comprehensive refactoring of the benchmark suite to fix critical bugs,
eliminate redundancies, and follow Rust/Criterion best practices.
Reduced from 12 benchmark files to 9 well-organized files.

## Bug Fixes (Critical)
**persistence.rs - 3 critical bugs fixed:**
1. bench_wal_with_sync: Now actually uses DurabilityMode parameter
   - Replaced unused sync_enabled boolean with DurabilityMode enum
   - Tests Async vs Synchronous modes with commit_with_mode()
   - Fixes issue where sync_on showed 477ns (impossible for fsync)

2. bench_wal_throughput: Now uses iter_batched for proper measurement
   - Moved TempDir/WAL creation to setup closure (not measured)
   - Only WAL append operations are timed
   - Fixes measurement of allocation overhead instead of throughput

3. Deleted bench_checkpoint_frequency (measured wrong thing)
   - Function measured should_checkpoint() predicate, not performance
   - Provided no useful performance data

## File Consolidation (12 → 9 files)
**Deleted files (merged into others):**
- benches/persistence.rs → split into durability_modes.rs + checkpoints.rs
- benches/id_generation.rs → merged into current_state.rs
- benches/string_interning.rs → merged into current_state.rs
- benches/concurrent_reads.rs → merged into temporal_query.rs

**New file:**
- benches/checkpoints.rs - checkpoint & recovery benchmarks

**Final 9 benchmark files:**
1. current_state.rs - Storage primitives (nodes, edges, ID gen, interning)
2. gallifreydb.rs - High-level API benchmarks
3. durability_modes.rs - WAL durability modes + low-level WAL operations
4. checkpoints.rs - Checkpoint creation, loading, recovery
5. transactions.rs - Transaction overhead and ACID guarantees
6. temporal_query.rs - Temporal indexes + concurrent access patterns
7. temporal_vector.rs - Temporal vector search
8. vector_similarity.rs - Vector similarity metrics
9. hnsw_index.rs - HNSW k-NN search

## Additional Fixes
- Fixed LSN subtraction (LSN.0 access for u64)
- Fixed type mismatches (usize vs u64 in assertions)
- Updated Cargo.toml with correct [[bench]] entries
- All benchmarks compile and pass clippy

## Verification
- ✅ All 9 benchmarks compile: cargo bench --no-run
- ✅ Clippy passes: cargo clippy --benches -- -D warnings
- ✅ Code formatted: cargo fmt --all
